chore: converted Box component to TS
